requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
PyMySQL==1.1.0
python-dotenv==1.0.0
werkzeug==3.0.1
email-validator==2.1.0
cryptography==41.0.7
